+++ src/plone.server/plone/server/cookiecutter/application/{{cookiecutter.package_name}}/{{cookiecutter.package_name}}/__init__.py
@@ -1,0 +1,10 @@
+app_settings = {
+    # provide custom application settings here...
+}
+
+
+def includeme(root):
+    """
+    custom application initialization here
+    """
+    pass