+++ src/plone.server/plone/server/cookiecutter/application/{{cookiecutter.package_name}}/{{cookiecutter.package_name}}/api.py
@@ -1,0 +1,10 @@
+from plone.server.api.service import Service
+from plone.server.browser import Response
+
+
+class ExampleService(Service):
+
+    async def __call__(self):
+        return Response({
+            'foo': 'bar'
+        })